New improves the cost engine.
1.-Now Allow delayed entry transaction, and the register the cost adjustment
2.-Add the IDocumentLine the AcctDate field and C_DocType necessary validate if period is open
3.-Create new method to get Cost detail based in Cost Dimension
4.-Create new constructor to create MCostDetail instance based in Cost Dimension, the cost dimension allow validate the Cost Level defined in Account Schema or Product Category 
5.-Enable the documents reference id to allow the update in application dictionary.
6.-New process to regenerate the CostDetail history from MTransaction history, it process will use to migrate from old version.
7.-New Average Cost test class

Business case:

Created a Purchase Order by 10 Oak with the price of 36.
Created a Material Receipt by 10 
Created a Customer Shipment by 5
Created a Invoice Vendor with the price of 38
Validate that the Cost Adjustment is working.
Validate that the Inventory Valued was Adjustment
Created a Material Receipt reverse.

kind regards
Victor Perez
www.e-evolution.com