1.- Generate cost detail for reverse transaction using the store engine logic.
Current Adempiere generate a transaction for unique ASI but what a reverse is generate use ASI 0, So I think that a reverse transaction should be create using the original ASI.

Is necessary review store engine and try fix this in the future, but to keep the back compatibility the cost engine use the same logic to generate the reverse.

For example: I generate the out inventory transaction for 10 each, the store engine generate 2 transaction the first by 3 with ASI 1000000 and the second by 7 with ASI 1000001, it is ok the issue is that when I generate a reverse  the transaction generate a single transaction by 10 with ASI 0.

2.- Fix Cost Valuation Effective Date
 
3.- expected division between zero

kind regards
Victor Perez
www.e-evolution.com