"Transaction Valuation" not calculating correctly B.Balance
The Physical Inventory use to staring the quantity on hand and cost via Import Inventory, the current approach the overwrite the cost into the MCost it can damage the cost calculate based on history transaction.

The new approach the cost is imported and save into the MInventoryLine a new field called currentcostprice where is save the cost from import table.

So the logic apply now is if a cost detail is found then this cost will use to create the inventory transaction cost, if a cost not found then use the current cost price reference that was imported. so that the import cost is copy and save inventory line table  from import table you can drop the import records without some issue.

kind regards
Victor Perez
www.e-evolution.com